+++ FrontEnd/src/components/HouseBrowsing/HouseBrowsing.tsx
@@ -6,19 +6,11 @@ import HouseCardList from '@components/HouseCardList';
 import styles from './HouseBrowsing.module.scss';
 
 const Housing: FunctionComponent = () => {
-  const showNewUserPopup = useShowNewUserPopup();
-
+  
   return (
     <>
       {/* TODO not currently used in this version... <HousingPost show={showHousingPost} setShow={setShowHousingPost} /> */}
 
-      {showNewUserPopup !== undefined && ( // only render the modal when user info exists, to initialize the wizard form with the user info
-        <NewUserSetup
-          show={showNewUserPopup !== undefined}
-          name={showNewUserPopup?.name}
-          email={showNewUserPopup?.email}
-        />
-      )}
 
       <div className={styles.filter}>
         <Filter />
